requirements.txt
pytest~=7.4.3
allure-pytest
selene>=2.0.0rc4
Appium-Python-Client~=3.1.0
python-dotenv==1.0.0
pydantic~=2.4.2
pydantic-settings~=2.0.3